+++ lib/TaskRouterClient.js
@@ -1,8 +1,8 @@
 /**
- @module WdsClient
+ @module TaskRouterClient
 
  This module presents a higher-level API for interacting with resources
- in the Twilio WDS API.  Tries to map very closely to the resource structure
+ in the Twilio TaskRouter API.  Tries to map very closely to the resource structure
  of the actual Twilio API, while still providing a nice JavaScript interface.
  */
 
@@ -12,15 +12,15 @@ var Client = require('./Client');
 var util = require('util');
 
 /**
- The Twilio WDS API client
+ The Twilio TaskRouter API client
  @constructor
  @param {string} sid - The application SID, as seen in the Twilio portal
  @param {string} tkn - The auth token, as seen in the Twilio portal
  @param {object} options (optional) - optional config for the REST client
- - @member {string} host - host for the Twilio API (default: wds.twilio.com)
+ - @member {string} host - host for the Twilio API (default: taskrouter.twilio.com)
  - @member {string} apiVersion - the Twilio REST API version to use for requests (default: v1)
  */
-function WdsClient(sid, tkn, workspaceSid, options) {
+function TaskRouterClient(sid, tkn, workspaceSid, options) {
     //Required client config
     if (!workspaceSid) {
         if (process.env.TWILIO_WORKSPACE_SID) {
@@ -35,16 +35,16 @@ function WdsClient(sid, tkn, workspaceSid, options) {
         this.workspaceSid = workspaceSid.replace(/ /g, '');
     }
     options = options || {};
-    WdsClient.super_.call(this, sid, tkn, options.host || 'wds.twilio.com', options.apiVersion || 'v1', options.timeout);
+    TaskRouterClient.super_.call(this, sid, tkn, options.host || 'taskrouter.twilio.com', options.apiVersion || 'v1', options.timeout);
 
     //REST Resource - shorthand for just "workspace" and "workspaces" to match the REST API
-    var workspaceResource = require('./resources/wds/Workspaces')(this);
+    var workspaceResource = require('./resources/task_router/Workspaces')(this);
     this.workspaces = workspaceResource;
 
     //mix the account object in with the client object - assume master account for resources
     _.extend(this, workspaceResource);
 }
 
-util.inherits(WdsClient, Client);
+util.inherits(TaskRouterClient, Client);
 
-module.exports = WdsClient;
+module.exports = TaskRouterClient;
