+++ lib/index.js
@@ -8,7 +8,7 @@
 
 var webhooks = require('./webhooks'),
     RestClient = require('./RestClient'),
-    WdsClient = require('./WdsClient');
+    TaskRouterClient = require('./TaskRouterClient');
 
 //Shorthand to automatically create a RestClient
 function initializer(sid, tkn, options) {
@@ -17,7 +17,7 @@ function initializer(sid, tkn, options) {
 
 //Main functional components of the Twilio module
 initializer.RestClient = RestClient;
-initializer.WdsClient = WdsClient;
+initializer.TaskRouterClient = TaskRouterClient;
 initializer.Capability = require('./Capability');
 initializer.TwimlResponse = require('./TwimlResponse');
