+++ lib/resources/task_router/Workspaces.js
@@ -1,6 +1,6 @@
 /**
- @module resources/wds/Workspaces
- The Twilio WDS "Workspaces" Resource.
+ @module resources/task_router/Workspaces
+ The Twilio TaskRouter "Workspaces" Resource.
  */
 var generate = require('../generate');
