+++ spec/client.activity.spec.js
@@ -1,7 +1,7 @@
 var twilio = require('../index');
 
-describe('The Twilio WDS Activity resource', function () {
-    var client = new twilio.WdsClient('AC123', '123', 'WS123');
+describe('The Twilio TaskRouter Activity resource', function () {
+    var client = new twilio.TaskRouterClient('AC123', '123', 'WS123');
 
     beforeEach(function () {
         spyOn(client, 'request');
